refactor(Modals): move Modals and associated components into separate directories
