+++ components/modals/Settings/Settings.tsx
@@ -11,11 +11,11 @@ import {
     Typography,
 } from "@mui/material";
 import React, { Dispatch, SetStateAction, useEffect, useState } from "react";
-import { ConfigPayload, NotificationsConfig } from "../types/rezervoTypes";
+import { ConfigPayload, NotificationsConfig } from "../../../types/rezervoTypes";
 import NotificationsActiveRoundedIcon from "@mui/icons-material/NotificationsActiveRounded";
 import PlayCircleOutlineRoundedIcon from "@mui/icons-material/PlayCircleOutlineRounded";
-import { DEFAULT_REMINDER_HOURS } from "../config/config";
-import { useUserConfig } from "../hooks/useUserConfig";
+import { DEFAULT_REMINDER_HOURS } from "../../../config/config";
+import { useUserConfig } from "../../../hooks/useUserConfig";
 
 // Fix track not visible with "system" color scheme
 const Switch = styled(MaterialUISwitch)(({ theme }) => ({
